refactor: reorganizar estructura modular
