+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from 'next'
 import { Inter, Poppins } from 'next/font/google'
 import './globals.css'
-import { AuthProvider } from '@/contexts/AuthContext'
+import { AuthProvider } from '@/features/auth'
 import { PerformanceWidget } from '@/components/ui/PerformanceWidget'
 
 const inter = Inter({
