+++ src/app/page.tsx
@@ -1,10 +1,10 @@
 'use client'
 
 import { useRouter } from 'next/navigation'
-import { Hero } from '@/components/features/hero'
-import { BookingButtons } from '@/components/features/booking'
-import { Reviews } from '@/components/features/reviews'
-import { SurfPlans } from '@/components/features/surf-plans'
+import { Hero } from '@/features/hero'
+import { BookingButtons } from '@/features/booking'
+import { Reviews } from '@/features/reviews'
+import { SurfPlans } from '@/features/surf-plans'
 import Header from '@/components/layout/Header'
 import Footer from '@/components/layout/Footer'
 import { ROUTES } from '@/lib/constants'
