--- src/components/features/booking/index.ts
@@ -1,1 +1,0 @@
-export { BookingButtons } from './BookingButtons'
