--- src/components/features/hero/index.ts
@@ -1,1 +1,0 @@
-export { Hero } from './Hero'
