--- src/components/features/surf-plans/index.ts
@@ -1,1 +1,0 @@
-export { default as SurfPlans } from './SurfPlans'
