+++ src/components/layout/Header/Header.tsx
@@ -4,7 +4,7 @@ import { useState } from 'react'
 import Link from 'next/link'
 import { Menu, X, Waves, User, LogOut } from 'lucide-react'
 import { motion, AnimatePresence } from 'framer-motion'
-import { useAuth } from '@/contexts/AuthContext'
+import { useAuth } from '@/features/auth/hooks/useAuth'
 import { Button } from '@/components/ui/Button'
 import { useResponsive } from '@/hooks/useResponsive'
 import { ANIMATION_DEFAULTS, APP_CONFIG } from '@/lib/constants'
