+++ src/features/auth/index.ts
@@ -1,0 +1,2 @@
+export { AuthProvider } from './components/AuthProvider'
+export { useAuth } from './hooks/useAuth'
