+++ src/features/booking/index.ts
@@ -1,0 +1,1 @@
+export { BookingButtons } from './components/BookingButtons'
