+++ src/features/catalog/services/.gitkeep
@@ -1,0 +1,1 @@
+// Archivo de marcador para mantener el directorio
