+++ src/features/surf-plans/index.ts
@@ -1,0 +1,1 @@
+export { default as SurfPlans } from './components/SurfPlans'
