+++ src/services/clerk/index.ts
@@ -1,0 +1,3 @@
+// Servicio de autenticación con Clerk
+// Se implementará en futuras iteraciones
+export {};
