+++ src/services/stripe/index.ts
@@ -1,0 +1,3 @@
+// Servicio de pagos con Stripe
+// Implementación pendiente
+export {};
